+++ lib/tinkoff_acquiring_ui.dart
@@ -1,7 +1,0 @@
-/// Содержит интерфейс, необходимый для подтверждения 3D Secure.
-///
-/// Основной класс модуля: `WebView3DS`
-library tinkoff_acquiring_ui;
-
-export 'src/ui/web_view_3ds.dart';
-export 'src/ui/collect_data.dart';