+++ packages/tinkoff_acquiring/test_components/sdk_single_test.dart
@@ -1,6 +1,6 @@
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 
-import 'package:tinkoff_acquiring/tinkoff_acquiring_core.dart';
+import 'package:tinkoff_acquiring/tinkoff_acquiring.dart';
 
 import 'sdk_test_constant.dart';