+++ packages/tinkoff_acquiring_flutter/lib/tinkoff_acquiring_flutter.dart
@@ -1,0 +1,4 @@
+library tinkoff_acquiring_flutter;
+
+export 'src/web_view_3ds.dart';
+export 'src/collect_data.dart';