+++ packages/tinkoff_acquiring/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 3.0.0
+
+* **BREAKING**: Splitting the library into two packages:
+  - tinkoff_acquiring: Package for working with Tinkoff API
+  - tinkoff_acquiring_flutter: A package for working with UI, at the moment it is WebView. In the future, it will be QR and much more
+
 # 2.1.0
 
 * Add Base Model & Update models